# Conflicts:
#	GolemEditor/Saves/Scenes/scene_0.json
#	GolemEngine/Include/Resource/Rendering/Shader/computeShader.h
#	GolemEngine/Include/Resource/resourceManager.h
#	GolemEngine/Include/WorldBuilder/terrain.h
#	GolemEngine/Include/Wrappers/graphicWrapper.h
#	GolemEngine/Source/Resource/resourceManager.cpp

+++ GolemEngine/Source/Components/audio.cpp
@@ -111,7 +111,7 @@ void Audio::SetLoop(bool _isLooping)
 
 void Audio::SetVolume(int _volume)
 {
-    m_volume = _volume;
+    m_volume = (float)_volume;
 }
 
 void Audio::StopMusic(bool _isPlaying)
